refactor(SCRUM-385): move EventCard.tsx to components folder
